Refactor Peg
