+++ ui/core/src/usecases/peg/utils/isOriginallySifchainNativeToken.ts
@@ -1,0 +1,5 @@
+import { IAsset } from "../../../entities";
+
+export function isOriginallySifchainNativeToken(asset: IAsset) {
+  return ["erowan", "rowan"].includes(asset.symbol);
+}
